Move non-keyservice system creation to onboarding step
