+++ server/portal/apps/auth/unit_test.py
@@ -1,15 +1,11 @@
 from django.test import (
-    TestCase, 
     TransactionTestCase,
-    Client, 
-    RequestFactory, 
     override_settings
 )
 from django.contrib.auth import get_user_model
-from mock import Mock, patch, MagicMock, ANY
+from mock import patch, MagicMock
 from portal.apps.auth.backends import AgaveOAuthBackend
 from requests import Response
-from portal.apps.accounts.models import PortalProfile
 from portal.apps.auth.views import launch_setup_checks
 import pytest
 
@@ -19,13 +15,9 @@ pytestmark = pytest.mark.django_db
 def test_launch_setup_checks(mocker, regular_user, settings):
     mocker.patch('portal.apps.auth.views.new_user_setup_check')
     mock_execute = mocker.patch('portal.apps.auth.views.execute_setup_steps')
-    mock_index = mocker.patch('portal.apps.auth.views.index_allocations')
-    mock_get_systems = mocker.patch('portal.apps.auth.views.get_user_storage_systems')
-    mock_get_systems.return_value = []
     regular_user.profile.setup_complete = False
     launch_setup_checks(regular_user)
     mock_execute.apply_async.assert_called_with(args=['username'])
-    mock_index.apply_async.assert_called_with(args=['username'])
 
 
 class TestAgaveOAuthBackend(TransactionTestCase):
@@ -55,10 +47,10 @@ class TestAgaveOAuthBackend(TransactionTestCase):
         # Test that backend failure responses are handled
 
         # Mock different return values for the backend response
-        self.mock_response.json.return_value = { }
+        self.mock_response.json.return_value = {}
         result = self.backend.authenticate(backend='agave', token='1234')
         self.assertIsNone(result)
-        self.mock_response.json.return_value = { "status" : "failure" }
+        self.mock_response.json.return_value = {"status": "failure"}
         result = self.backend.authenticate(backend='agave', token='1234')
         self.assertIsNone(result)
 
@@ -66,17 +58,17 @@ class TestAgaveOAuthBackend(TransactionTestCase):
     def test_new_user(self):
         # Test that a new user is created and returned
         self.mock_response.json.return_value = {
-            "status" : "success",
-            "result" : {
-                "username" : "testuser",
-                "first_name" : "test",
-                "last_name" : "user",
-                "email" : "test@user.com"
+            "status": "success",
+            "result": {
+                "username": "testuser",
+                "first_name": "test",
+                "last_name": "user",
+                "email": "test@user.com"
             }
         }
         result = self.backend.authenticate(backend='agave', token='1234')
         self.assertEqual(result.username, "testuser")
-    
+
     @override_settings(PORTAL_USER_ACCOUNT_SETUP_STEPS=[])
     def test_update_existing_user(self):
         # Test that an existing user's information is
@@ -91,12 +83,12 @@ class TestAgaveOAuthBackend(TransactionTestCase):
         )
         user.save()
         self.mock_response.json.return_value = {
-            "status" : "success",
-            "result" : {
-                "username" : "testuser",
-                "first_name" : "test",
-                "last_name" : "user",
-                "email" : "test@user.com"
+            "status": "success",
+            "result": {
+                "username": "testuser",
+                "first_name": "test",
+                "last_name": "user",
+                "email": "test@user.com"
             }
         }
         result = self.backend.authenticate(backend='agave', token='1234')
