+++ server/portal/apps/auth/views.py
@@ -12,10 +12,7 @@ from django.urls import reverse
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponseRedirect, HttpResponseBadRequest
 from django.shortcuts import render
-from celery import group
 from portal.apps.auth.models import AgaveOAuthToken
-from portal.apps.auth.tasks import setup_user, get_user_storage_systems
-from portal.apps.search.tasks import index_allocations
 from portal.apps.onboarding.execute import (
     execute_setup_steps,
     new_user_setup_check
@@ -67,15 +64,6 @@ def launch_setup_checks(user):
     if not user.profile.setup_complete:
         logger.info("Executing onboarding setup steps for %s", user.username)
         execute_setup_steps.apply_async(args=[user.username])
-
-    # Apply asynchronous long calls
-    logger.info("Starting system setup celery tasks for {username}".format(username=user.username))
-    index_allocations.apply_async(args=[user.username])
-
-    system_names = get_user_storage_systems(user.username, settings.PORTAL_DATA_DEPOT_LOCAL_STORAGE_SYSTEMS)
-    group(
-        setup_user.s(user.username, system) for system in system_names
-    ).apply_async()
 
 
 def agave_oauth_callback(request):
